[FEAT] 카테고리 repository 구현(삽입, 삭제, 조회). 메모 가져올 때 카테고리 가져오는 로직 수정 (categoryDao 를 통해 가져옴). 메모 업로드 시 카테고리 Spinner 목록 Room 에서 가져오도록 구현.

+++ domain/src/main/java/com/jae464/domain/repository/MemoRepository.kt
@@ -17,7 +17,7 @@ interface MemoRepository {
     fun getAllMemoWithPage(): Flow<PagingData<Memo>>
     suspend fun getAllMemoByUserOnRemote(uid: String): List<Memo>
 
-    fun getMemoByCategory(category: Int) : Flow<List<Memo>>
+    fun getMemoByCategory(category: Long) : Flow<List<Memo>>
     fun getMemoByTitle(title: String): Flow<List<Memo>>
     fun getMemoByContent(content: String): Flow<List<Memo>>
